Please enter the commit message for your changes. Lines starting with '#' will be ignored, and an empty message aborts the commit.
On branch master
Your branch is up to date with 'origin/master'.

Changes to be committed:
	renamed:    build/.cmake/api/v1/reply/codemodel-v2-93726edac95edfe6e7a6.json -> build/.cmake/api/v1/reply/codemodel-v2-d5abe3e7050b9c292d58.json
	renamed:    build/.cmake/api/v1/reply/index-2023-11-10T12-31-36-0057.json -> build/.cmake/api/v1/reply/index-2023-11-11T09-42-11-0837.json
	renamed:    build/.cmake/api/v1/reply/target-api-Debug-4883e450d84d56b73abe.json -> build/.cmake/api/v1/reply/target-api-Debug-dbba3192e975d61cc7a0.json
	modified:   build/CMakeFiles/Makefile2
	modified:   build/CMakeFiles/progress.marks
	modified:   build/compile_commands.json
	modified:   build/src/CMakeFiles/atsys.dir/CXX.includecache
	modified:   build/src/CMakeFiles/atsys.dir/depend.internal
	modified:   build/src/CMakeFiles/atsys.dir/depend.make
	modified:   build/src/CMakeFiles/atsys.dir/main.cc.o
	modified:   build/src/CMakeFiles/atsys.dir/progress.make
	modified:   build/src/CMakeFiles/progress.marks
	modified:   build/src/api/CMakeFiles/api.dir/CXX.includecache
	modified:   build/src/api/CMakeFiles/api.dir/DependInfo.cmake
	modified:   build/src/api/CMakeFiles/api.dir/api_manager.cc.o
	modified:   build/src/api/CMakeFiles/api.dir/build.make
	modified:   build/src/api/CMakeFiles/api.dir/card_api.cc.o
	modified:   build/src/api/CMakeFiles/api.dir/cmake_clean.cmake
	modified:   build/src/api/CMakeFiles/api.dir/data_api.cc.o
	modified:   build/src/api/CMakeFiles/api.dir/depend.internal
	modified:   build/src/api/CMakeFiles/api.dir/depend.make
	modified:   build/src/api/CMakeFiles/api.dir/link.txt
	modified:   build/src/api/CMakeFiles/api.dir/login_api.cc.o
	modified:   build/src/api/CMakeFiles/api.dir/logout_api.cc.o
	modified:   build/src/api/CMakeFiles/api.dir/progress.make
	modified:   build/src/api/CMakeFiles/api.dir/register_api.cc.o
	new file:   build/src/api/CMakeFiles/api.dir/time_api.cc.o
	modified:   build/src/api/CMakeFiles/api.dir/user_api.cc.o
	modified:   build/src/api/CMakeFiles/progress.marks
	modified:   build/src/api/Makefile
	modified:   build/src/api/libapi.a
	modified:   build/src/app/CMakeFiles/app.dir/CXX.includecache
	modified:   build/src/app/CMakeFiles/app.dir/app.cc.o
	modified:   build/src/app/CMakeFiles/app.dir/depend.internal
	modified:   build/src/app/CMakeFiles/app.dir/depend.make
	modified:   build/src/app/CMakeFiles/app.dir/progress.make
	modified:   build/src/app/libapp.a
	modified:   build/src/atsys
	modified:   build/src/base/CMakeFiles/base.dir/CXX.includecache
	modified:   build/src/base/CMakeFiles/base.dir/CachePool.cc.o
	modified:   build/src/base/CMakeFiles/base.dir/progress.make
	modified:   build/src/base/libbase.a
	modified:   build/src/global/CMakeFiles/global.dir/CXX.includecache
	modified:   build/src/global/CMakeFiles/global.dir/depend.internal
	modified:   build/src/global/CMakeFiles/global.dir/depend.make
	modified:   build/src/global/CMakeFiles/global.dir/init.cc.o
	modified:   build/src/global/CMakeFiles/global.dir/progress.make
	modified:   build/src/global/CMakeFiles/progress.marks
	modified:   build/src/global/libglobal.a
	modified:   build/src/layer/CMakeFiles/layer.dir/CXX.includecache
	modified:   build/src/layer/CMakeFiles/layer.dir/depend.internal
	modified:   build/src/layer/CMakeFiles/layer.dir/depend.make
	modified:   build/src/layer/CMakeFiles/layer.dir/event_layer.cc.o
	modified:   build/src/layer/CMakeFiles/layer.dir/net_layer.cc.o
	modified:   build/src/layer/CMakeFiles/layer.dir/progress.make
	modified:   build/src/layer/CMakeFiles/progress.marks
	modified:   build/src/layer/liblayer.a
	modified:   doc/api.md
	modified:   include/api/api_manager.h
	modified:   include/api/context.h
	new file:   include/api/time_api.h
	modified:   include/base/cachepool/CachePool.h
	modified:   include/global/global.h
	modified:   include/global/init.h
	modified:   include/layer/event_layer.h
	modified:   sbin/atsys
	modified:   sql/web.sql
	modified:   src/api/logout_api.cc
	new file:   src/api/time_api.cc
	modified:   src/base/CachePool.cc
	modified:   src/global/init.cc
	modified:   src/layer/event_layer.cc
	modified:   src/main.cc

+++ include/base/cachepool/CachePool.h
@@ -49,7 +49,7 @@ public:
 	// ------------------- 字符串相关 -------------------
 	string get(string key);
 	string set(string key, string &value);
-	string setex(string key, int timeout, string value);
+	string setex(string key, int64_t timeout, string value);
 
 	// string mset(string key, map);
 	// 批量获取
